Move component files under its directory
