+++ src/components/BannerMenu/index.tsx
@@ -1,0 +1,1 @@
+export { BannerMenu } from "./BannerMenu";
