+++ src/components/Button/index.tsx
@@ -1,0 +1,2 @@
+export { Button, TransparentButton, HoverButton } from "./Button";
+export type { ButtonProps, TransparentButtonProps } from "./Button";
