+++ src/components/IconButton/IconButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { IconType } from "react-icons/lib";
-import { FontLevel, fontLevel2FontSize } from "../types/style";
-import { ButtonProps, HoverButton, TransparentButton } from "./Button";
+import { FontLevel, fontLevel2FontSize } from "../../types/style";
+import { ButtonProps, HoverButton, TransparentButton } from "../Button";
 
 type IconButtonProps = {
   children?: React.ReactNode;
