+++ src/components/IconInput/index.tsx
@@ -1,0 +1,1 @@
+export { IconInput } from "./IconInput";
